refactor out generate_code_for_parse
